- calc.cpp file is complete - add Calculator and Parser classes

+++ Calculator/Calculatable.h
@@ -4,6 +4,6 @@
 class Calculatable
 {
 public:
-	virtual int calculate() = 0;
+	virtual double calculate() = 0;
 };
